Refactor mapper, handler, fixtures and getBattlesInfo tests
