+++ src/battle/dto/mappers.ts
@@ -2,15 +2,19 @@ import type { Battle } from "../../types";
 import type { BattleDto } from "./types";
 
 export const mapBattleDtoToBattle = (battleDto: BattleDto): Battle => {
+  const isImageMissing =
+    battleDto.imageUrl ??
+    "https://static.wikia.nocookie.net/starwars/images/e/ed/Binks22BBY.png";
+
+  const doesImageNeedAlt = battleDto.imageUrl
+    ? `General view of ${battleDto.name}`
+    : "";
+
   return {
     ...battleDto,
     id: battleDto._id,
-    imageUrl: !battleDto.imageUrl
-      ? "https://static.wikia.nocookie.net/starwars/images/3/34/Spacebattle.jpg"
-      : battleDto.imageUrl,
-    imageAlt: !battleDto.imageUrl
-      ? `General view of a space battle over Coruscant due to missing image for ${battleDto.name}`
-      : `General view of ${battleDto.name}`,
+    imageUrl: isImageMissing,
+    imageAlt: doesImageNeedAlt,
   };
 };
